Tracking number of explored nodes
